+++ src/__tests__/unit/utils.test.ts
@@ -92,6 +92,7 @@ describe("buildPluginConfig", () => {
       },
       npmPublishRegistryConfig: {
         awsProfile: "aws-profile-2",
+        preferAwsEnvironmentCredentials: "true",
       },
     });
   });