+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.6.0] - 2020-12-28
 [0.6.0]: https://github.com/mhassan1/yarn-plugin-aws-codeartifact/compare/v0.5.0...v0.6.0
 
+- Add support for `preferAwsEnvironmentCredentials` in `PluginRegistryConfig`
 - Bump dev dependencies
 
 ## [0.5.0] - 2020-12-23